# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the Sober app",
        "title": "Sober API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/analytics/timeline/bac": {
            "get": {
                "description": "Calculate BAC for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get BAC calculation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start time",
                        "name": "start_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End time",
                        "name": "end_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Weight in kg",
                        "name": "weight_kg",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Gender",
                        "name": "gender",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Time step in minutes",
                        "name": "time_step_mins",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.BACCalculationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticate a user and generate a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login a user",
                "parameters": [
                    {
                        "description": "User login request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "description": "Retrieve the current user's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get the current user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserMeResponse"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Create a new user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign up a new user",
                "parameters": [
                    {
                        "description": "User signup request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UserSignupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserSignupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Similar to the http status code",
                    "type": "integer",
                    "x-order": "1",
                    "example": 400
                },
                "type": {
                    "description": "The type of error",
                    "type": "string",
                    "enum": [
                        "validation",
                        "database",
                        "entity"
                    ],
                    "x-order": "2",
                    "example": "validation"
                },
                "message": {
                    "description": "A human-readable error message",
                    "type": "string",
                    "x-order": "3",
                    "example": "Invalid request Body"
                },
                "details": {
                    "description": "Additional details about the error, omitted if empty",
                    "type": "array",
                    "items": {
                        "type": "object"
                    },
                    "x-order": "4"
                },
                "correlation_id": {
                    "description": "The error id from context traceId",
                    "type": "string",
                    "x-order": "5",
                    "example": "01234567890123456789012345678900"
                }
            }
        },
        "dtos.BACCalculationResponse": {
            "type": "object",
            "properties": {
                "summary": {
                    "$ref": "#/definitions/models.BACSummary"
                },
                "timeline": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BACPoint"
                    }
                }
            }
        },
...
        "dtos.UserLoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dtos.UserLoginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "dtos.UserMeResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "dtos.UserSignupRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dtos.UserSignupResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.BACPoint": {
            "type": "object",
            "properties": {
                "bac": {
                    "type": "number"
                },
                "is_over_bac": {
                    "type": "boolean"
                },
                "status": {
                    "$ref": "#/definitions/models.BACStatus"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "models.BACStatus": {
            "type": "string",
            "enum": [
                "Sober",
                "Minimal",
                "Light",
                "Mild",
                "Significant",
                "Severe",
                "Dangerous"
            ],
            "x-enum-varnames": [
                "BACStatusSober",
                "BACStatusMinimal",
                "BACStatusLight",
                "BACStatusMild",
                "BACStatusSignificant",
                "BACStatusSevere",
                "BACStatusDangerous"
            ]
        },
        "models.BACSummary": {
            "type": "object",
            "properties": {
                "drinking_since_time": {
                    "type": "string"
                },
                "duration_over_bac": {
                    "type": "integer"
                },
                "max_bac": {
                    "type": "number"
                },
                "max_bac_time": {
                    "type": "string"
                },
                "sober_since_time": {
                    "type": "string"
                },
                "total_drinks": {
                    "type": "integer"
                }
            }
        },
...
    }
}